+++ ui/public/theme/base.css
@@ -114,14 +114,9 @@ html {
     box-shadow: 0px 2px 1px -1px rgba(0, 0, 0, 0.2), 0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
     background-color: #fff;
     color: rgba(0, 0, 0, 0.87);
-    -webkit-transition: box-shadow 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
     transition: box-shadow 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-    border-radius: 4px;
-    box-shadow: var(--Paper-shadow);
-    background-image: var(--Paper-overlay);
-    overflow: hidden;
-    width: 100%;
-    overflow-x: auto;
+    border-radius: 10px;
+        margin: 15px;
 }
 
 .NuoTableTable {
@@ -147,6 +142,7 @@ html {
     outline: 0;
     unicode-bidi: isolate;
     border-color: inherit;
+    border-top: 1px solid rgba(224, 224, 224, 1);
 }
 
 .NuoTableSettingsItem {
@@ -179,12 +175,19 @@ html {
     display: table-cell;
     padding: 16px;
     vertical-align: inherit;
-    border-bottom: 1px solid rgba(224, 224, 224, 1);
     color: rgba(0, 0, 0, 0.87);
     unicode-bidi: isolate;
     text-align: left;
 }
 
+.NuoTableTh {
+    background-color: #f5f6f8;
+}
+
+.NuoTableTd {
+    background-color: white;
+    opacity: .8;
+}
 .NuoTableTd>button {
     border: none;
     background-color: transparent;
@@ -214,8 +217,6 @@ html {
     position: sticky;
     right: 0;
     text-align: right;
-    background-color: rgba(255, 255, 255, 0.8);
-    ;
 }
 .NuoCrontabContainer {
     display: flex;
@@ -540,4 +541,19 @@ code {
     display: flex;
     flex-direction: row;
     justify-content: space-between;
+}
+.NuoTableField_labels {
+    display: inline-block;
+    border: 1px solid lightgray;
+    border-radius: 99999px;
+    padding: 2px 8px;
+    margin: 2px;
+}
+
+*:has(>.NuoTableField_labels) {
+    display: inline-block;
+}
+
+.NuoTableOptions {
+    padding: 15px;
 }
