+++ ui/src/components/pages/parts/Path.tsx
@@ -1,55 +1,17 @@
 // (C) Copyright 2024-2025 Dassault Systemes SE.  All Rights Reserved.
 
-import React, { useState } from "react"
 import { withTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom"
-import Button from "../../controls/Button"
 import Breadcrumbs from '@mui/material/Breadcrumbs'
 import Select, { SelectOption } from "../../controls/Select"
 import Link from '@mui/material/Link'
 import Typography from '@mui/material/Typography';
-import TextField from "../../controls/TextField"
 import { styled } from '@mui/material';
 import { RestSpinner } from './Rest';
 import { getFilterField, getSchemaPath } from "../../../utils/schema";
 import { TempAny } from "../../../utils/types"
 
-export function parseSearch(search: string) {
-    search = search.trim();
-    let ret: TempAny = {};
-
-    search.split(" ").forEach((parts: string) => {
-        const posEqual = parts.indexOf("=");
-        if (posEqual !== -1) {
-            const key = parts.substring(0, posEqual);
-            const value = parts.substring(posEqual + 1);
-            if (key !== "name" && key !== "label") {
-                ret["error"] = "Invalid search key \"" + key + "\". Can only search for \"name\" or \"label\"";
-            }
-            else if (key in ret) {
-                ret["error"] = "Key \"" + key + "\" can only be specified once.";
-            }
-            else {
-                ret[key] = value;
-            }
-        }
-        else if (parts.indexOf(" ") !== -1) {
-            // ignore double spaces
-        }
-        else if ("name" in ret) {
-            ret["error"] = "Cannot search for \"name\" attribute multiple times";
-        }
-        else {
-            ret["name"] = parts;
-        }
-    });
-    return ret;
-}
-
-function Path({ schema, path, filterValues, search, setSearch, setPage, org, t }: TempAny) {
-    const [searchField, setSearchField] = useState(search);
-    const [error, setError] = useState(undefined);
-
+function Path({ schema, path, filterValues, org, t }: TempAny) {
     const navigate = useNavigate();
 
     const StyledBreadcrumbs = styled(Breadcrumbs)({
@@ -79,18 +41,6 @@ function Path({ schema, path, filterValues, search, setSearch, setPage, org, t }
             <SelectOption value="">{t("control.select.item.all")}</SelectOption>
             {filterValues && filterValues.map((fv: string) => <SelectOption key={fv} value={fv}>{fv}</SelectOption>)}
         </Select>;
-    }
-
-    function handleSearch() {
-        const parsed: TempAny = parseSearch(searchField);
-        if ("error" in parsed) {
-            setError(parsed["error"]);
-        }
-        else {
-            setError(undefined);
-            setSearch(searchField);
-            setPage(1);
-        }
     }
 
     let filterField = getFilterField(schema, path);
@@ -124,27 +74,6 @@ function Path({ schema, path, filterValues, search, setSearch, setPage, org, t }
             {renderFilter()}
         </StyledBreadcrumbs>
         <RestSpinner />
-        {setSearch && <React.Fragment>
-            <TextField
-                required={false}
-                data-testid="searchField"
-                id="search"
-                label={"search, i.e. \"somename\" or \"label=key=value,!otherkey name=abc\""}
-                value={searchField}
-                onChange={({ currentTarget: input }) => {
-                    setSearchField(input.value);
-                }}
-                onKeyDown={(event) => {
-                    if (event.keyCode === 13) {
-                        handleSearch();
-                    }
-                }}
-                error={error}
-            />
-            <Button data-testid="searchButton" onClick={handleSearch}
-            >{t("button.search")}</Button>
-        </React.Fragment>
-        }
     </div>;
 }
